User and MainMenu Update.
Playlists have been added, including their Title Instances.

A couple of Armor options have been partially added to the shop as I figure out how 'Armor_Item' works.

Users Updated In-Game Public Data are now stored in the data base, only thing left to do with this is to make it so the game uses that data which I will probably get to inshallah.