Shop and User Update.
(WI.P) Renting finally works, meaning users have the option to pay with Gold to Rent a Weapon or consumable for an hour (Credits don't work at the moment.), the API will update the time amount for every item you have rented, though a bit more work needs to be put into before it can be finished.

Weapons, Consumables and Armors now have their final prices for Renting and full purchase.

All Weapons can now be assigned to Loadouts instead of some not being able to be assigned to Loadout.

Users will automatically have Air Assault armor unlocked.

Please note that Setting any of the Weapon/Armor for the Main Menu wont work at the moment, as I will get to that soon, but in the meantime enjoy :+1: