User + Transactions + Shop + Menu Update
When logging in, The server can now detect when the User is online which makes it easier to invite the specified User.

Users can now finally pay with Credits (PROPERLY) as another form of currency.

Message of the day has been implemented, and also fixed the weird issue where you'd have to scroll down to see the news message.

Still a lot left to do, I'm now somewhat shifting my focus towards the challenges system which I've got an idea on how to implement it, Class Select screen still needs work and will work once I get armors to work.